move column-related things into adapters where they belong
