+++ core/dbt/adapters/base/__init__.py
@@ -1,6 +1,7 @@
 # these are all just exports, #noqa them so flake8 will be happy
 from dbt.adapters.base.meta import available  # noqa
 from dbt.adapters.base.relation import BaseRelation  # noqa
+from dbt.adapters.base.relation import Column  # noqa
 from dbt.adapters.base.connections import BaseConnectionManager  # noqa
 from dbt.adapters.base.connections import Credentials  # noqa
 from dbt.adapters.base.impl import BaseAdapter  # noqa
