+++ plugins/bigquery/dbt/adapters/bigquery/__init__.py
@@ -1,6 +1,7 @@
 from dbt.adapters.bigquery.connections import BigQueryConnectionManager  # noqa
 from dbt.adapters.bigquery.connections import BigQueryCredentials
 from dbt.adapters.bigquery.relation import BigQueryRelation  # noqa
+from dbt.adapters.bigquery.relation import BigQueryColumn  # noqa
 from dbt.adapters.bigquery.impl import BigQueryAdapter
 
 from dbt.adapters.base import AdapterPlugin
